Moved settings route to user route file due to filename conflict.
