+++ app.py
@@ -13,9 +13,8 @@ from lib.routes.patients import patch_intake_route, search_patients_route, patie
 from lib.routes.testing import test_surrealdb_route, test_crud_route, debug_session_route
 from lib.routes.users import search_users_route, check_users_exist_route, setup_default_admin_route, \
     activate_user_route, deactivate_user_route, get_all_users_route, change_password_route, get_current_user_info_route, \
-    logout_route, register_route, login_route
+    logout_route, register_route, login_route, settings_route
 from lib.services.auth_decorators import require_auth, require_admin, optional_auth
-from lib.routes.settings import settings_route
 from settings import PORT, DEBUG, HOST, FLASK_SECRET_KEY
 
 app = Flask(__name__)
